+++ back-end/pdf_dowloader.py
@@ -1,3 +1,4 @@
+#!/usr/bin/env python3
 import os
 import requests
 from bs4 import BeautifulSoup
